Fix: deserializing nested objects
Add ReflectionExtractor to get the type of nested fields from their types.
This also breaks GET payloads, since all of their types are "string", because they are transported using a query string. Without the ReflectionExtractor, the discovered type was `null` and the value was coalesced to whatever the class had. With it, the correct type is discovered, but a mismatch happens, because all input types are "string". Adding `csv` format when deserializing GET payloads triggers the type juggling/coalescing inside of the ObjectNormalizer and everything starts working again